Add files via upload
This work is to study the OSA which is a common sleep disorder, this disease causes a large population to not sleep well. So this study is to predict trough machine learning the level of IAH with regression and classify the patients if they are healthy or severe with classification.

So the first step was cleaning our data in order to keep only the interesting features and discard lines with missing values. Then through EDA, we have found the features that are highly correlated to IAH: weight, Cervical, BMI and Age. With these features we developed our models for regression and classification. After training the models we tested it and selected the best ones for each study: Lasso  model for regression and random forest model for classification. These are two that should be used to predict. 

This study was interesting and was helpful in order to understand the functioning of machine learning. Indeed, this work manage to help me to understand the importance of each step of machine learning. It let me to try the functions that I have learned during classes and develop my creativity in python.

In conclusion, this work allows both my personal development and gives interesting results for “Quirón Hospital of Málaga”.